requirements.txt
pytest-playwright
playwright~=1.51.0
beautifulsoup4~=4.13.4
Pillow
cairosvg
streamlit
python-dotenv
websockets